5th

+++ AccountUser.java
@@ -3,10 +3,10 @@ package bank;
 public class AccountUser extends Thread{
 
 	Account account;
-	int amount;
+	double amount;
 	boolean b;
 	
-	public AccountUser(Account account,int amount,boolean b) {
+	public AccountUser(Account account,double amount,boolean b) {
 		this.account=account;
 		this.amount=amount;
 		this.b=b;
@@ -14,13 +14,10 @@ public class AccountUser extends Thread{
 	
 	@Override
 	public void run() {
-		
-		//for(int i=0;i<4;i++) {
-			if(b==true) {
-				account.deposit(amount);
-			}else {
-				account.withdraw(amount);
-			}
-		//}
+		if(b==true) {
+			account.deposit(amount);
+		}else {
+			account.withdraw(amount);
+		}
 	}
 }
